extract addSongModal
